+++ src/sequential.rs
@@ -29,11 +29,11 @@ impl EvalGradient for Sequential {
             7 => self.seven[i],
             8 => self.eight[i],
             9 => self.nine[i],
-            _ => self.eval_continuous(i as f64 / (n - 1) as f64),
+            _ => self.eval_continuous(i as f32 / (n - 1) as f32),
         }
     }
 
-    fn eval_continuous(&self, t: f64) -> Color {
+    fn eval_continuous(&self, t: f32) -> Color {
         interpolate::spline(&self.nine, t)
     }
 }
